refactor!: Move expression transform code to its own module
